0.9.2 commit...
* Text on icon area of top screen graphics show "No Icon" instead of
original text of DS Cart. This text is only visible if no NDS file is
selected or an NDS file that is selected has no icon.
* Fix top screen text on load to display HBMenu instead of Acekard 2i
(this project was originally setup as a bootstrap for that cart when I
initially did this UI overhaul)
* Center text for Fat Init Fail error text for bottom screen.
* Added new line of text for bottom screen displayed during FAT init.
Only really visible if you have an especially slow card or if DLDI init
hangs. This replaces original HBMenu's extra text that used to be
displayed with it's name on top screen during init.